Struct sangfroid::threadpool::ThreadPool

pub struct ThreadPool<Req, Res> where
    Req: Send + Debug + 'static,
    Res: Send + Debug + 'static,  {
    pool: Option<Arc<Mutex<BinaryMaxHeap<Worker<Req, Res>>>>>,
    done_channel: Sender<Option<u64>>,
    balancer: Option<JoinHandle<Result<(), ThreadPoolError>>>,
}

ThreadPool to keep track of worker threads, dynamically dispatch jobs in a load balanced manner, and distribute load evenly.

Fields
pool: Option<Arc<Mutex<BinaryMaxHeap<Worker<Req, Res>>>>>done_channel: Sender<Option<u64>>balancer: Option<JoinHandle<Result<(), ThreadPoolError>>>
Implementations

impl<Req, Res> ThreadPool<Req, Res> where
Req: Send + Debug + 'static,
Res: Send + Debug + 'static,

pub fn new(workers: usize) -> Self

Creates a new threadpool with the given number of workers and the optionally provided Request and Result types.

use sangfroid::threadpool::ThreadPool;
use sangfroid::job::Job;

let thread_pool = ThreadPool::<u8, u8>::new(1);

let (job, result_src) = Job::with_result_sink(|x| x * 2, 2);
thread_pool.schedule(job).expect("job not scheduled");
assert_eq!(result_src.recv().unwrap(), 4);

pub fn balancer_thread(
done_channel: Receiver<Option<u64>>,
worker_heap: Arc<Mutex<BinaryMaxHeap<Worker<Req, Res>>>>
) -> JoinHandle<Result<(), ThreadPoolError>>

Returns a JoinHandle to a balancer thread for the given worker pool. The balancer listens on the given done receiver channel to receive Uids of workers who have finished their job and need to get their load decremented. The core loop of the balancer may be described as follows in pseudocode:

while uid = done_channel.recv() {
    restore_worrker_pool_order(worker_pool, uid)
}

Since the worker pool is shared with the main thread for dispatching jobs, we need to wrap it in a Mutex.

pub fn new_workers(
workers: usize
) -> (Vec<Worker<Req, Res>>, (Sender<Option<u64>>, Receiver<Option<u64>>))

Creates the given number of workers and returns them in a vector along with the ends of the done channel. The workers send their Uid to the send end of the done channel to signify completion of a job. The balancer thread received on the receiving end of the done channel for uids and balances them accordingly.

One of the key decisions behind this library is that we move channels where they are tobe used instead of sharing them with a lock. The send end of the channel is cloned and passed to each of the workers. The receiver end returned is meant to be moved to the balancer thread’s closure.

pub fn schedule(&self, job: Job<Req, Res>) -> Result<(), ThreadPoolError>

Schedules a new job to the worker pool by picking up the least loaded worker and dispatching the job to it.

use sangfroid::threadpool::ThreadPool;
use sangfroid::job::Job;

let thread_pool = ThreadPool::<u8, u8>::new(2);

let (job, result_src) = Job::with_result_sink(|x| x * 2, 2);
thread_pool.schedule(job).expect("job not scheduled");
assert_eq!(result_src.recv().unwrap(), 4);

let (job, result_src) = Job::with_result_sink(|x| x * 3, 2);
thread_pool.schedule(job).expect("job not scheduled");
assert_eq!(result_src.recv().unwrap(), 6);

let (job, result_src) = Job::with_result_sink(|x| x * 4, 2);
thread_pool.schedule(job).expect("job not scheduled");
assert_eq!(result_src.recv().unwrap(), 8);

pub fn terminate(&mut self) -> Result<(), ThreadPoolError>

Terminates this threadpool by invoking theadpool::worker_pool_terminated() on the worker pool and sending a None to the balancer thread via the done channek to notify termination. We then invoke join() on the balancer thread.

We keep resources with custom cleanup inside Options so as to be able to move them in to the current scrope and drop() them.

use sangfroid::threadpool::ThreadPool;
use sangfroid::job::Job;

let mut thread_pool = ThreadPool::<u8, u8>::new(2);
thread_pool.terminate();