Struct sangfroid::worker::Worker

pub struct Worker<Req, Res> where
    Req: Send + Debug + 'static,
    Res: Send + Debug + 'static,  {
    uid: u64,
    disp_q: Sender<Message<Req, Res>>,
    worker: Option<JoinHandle<Result<(), WorkerError>>>,
    pending: usize,
}

Worker represents a worker thread capable for receiving and servicing jobs.

Fields
uid: u64

uid for uniquely identifying this worker

disp_q: Sender<Message<Req, Res>>

message dispatch queue

worker: Option<JoinHandle<Result<(), WorkerError>>>

worker thread for executing jobs

pending: usize

number of pending jobs to be serviced

Implementations

impl<Req, Res> Worker<Req, Res> where
Req: Send + Debug + 'static,
Res: Send + Debug + 'static,

pub fn new(
job_source: Receiver<Message<Req, Res>>,
disp_q: Sender<Message<Req, Res>>,
done: Sender<Option<u64>>,
uid: u64
) -> Self

Creates a new Worker from the given source, dispatch queue channel and done notice channel. The done channel and job_source channel are moved into the worker thread closure for receiving requests and notifying done status respectively.

The worker expects the mpsc::Receiver for the done mpsc::Sender to outlive itself.

pub fn worker_thread(
jobs: Receiver<Message<Req, Res>>,
done: Sender<Option<u64>>,
uid: u64
) -> JoinHandle<Result<(), WorkerError>>

Creates a worker thread from the given job source, done notification channel and worker uid. This is not meant to be used directly. It is a advisable to construct a Worker instead since the Worker instance also manages the lifecycle and cleanup of the thread.

/// The worker thread core loop

// ...
while let Ok(Request(job)) = job_source.recv() {
    job.result_channel.send(job.task(job.req));
    done.send(worker_uid);
}
// ...

pub fn dispatch(&self, job: Job<Req, Res>) -> Result<(), WorkerError>

Dispatches a job to this worker for execution.

pub fn terminate(&mut self) -> Result<(), WorkerError>

Terminates this worker by sending a Terminate message to the underlying worker thread and the invoking join() on it.

pub fn inc_load(&mut self)

Increments pending tasks by 1.

pub fn dec_load(&mut self)

Decrements pending tasks by 1.